refactor(session-ui): move timeline v2 components
