+++ packages/app/src/pages/session/timeline/message-timeline.tsx
@@ -56,6 +56,10 @@ import { showToast } from "@/utils/toast"
 import { getDirectory, getFilename } from "@opencode-ai/core/util/path"
 import { Popover as KobaltePopover } from "@kobalte/core/popover"
 import { normalize } from "@opencode-ai/session-ui/session-diff"
+import { scheduleConnectedMeasure } from "@opencode-ai/session-ui/v2/timeline-measure-v2"
+import { observeElementOffsetReconnectAware } from "@opencode-ai/session-ui/v2/timeline-observe-element-offset-v2"
+import { createTimelineProjection } from "@opencode-ai/session-ui/v2/timeline-projection-v2"
+import { filterVirtualIndexes } from "@opencode-ai/session-ui/v2/timeline-virtual-items-v2"
 import { useFileComponent } from "@opencode-ai/ui/context/file"
 import { shouldMarkBoundaryGesture, normalizeWheelDelta } from "@/pages/session/message-gesture"
 import { SessionContextUsage } from "@/components/session-context-usage"
@@ -71,11 +75,7 @@ import { useSDK } from "@/context/sdk"
 import { useSync } from "@/context/sync"
 import { notifySessionTabsRemoved } from "@/components/titlebar-session-events"
 import { sessionTitle } from "@/utils/session-title"
-import { scheduleConnectedMeasure } from "./measure"
-import { observeElementOffsetReconnectAware } from "./observe-element-offset"
-import { createTimelineProjection } from "./projection"
-import { MessageComment, SummaryDiff, TimelineRow, TimelineRowMap } from "./rows"
-import { filterVirtualIndexes } from "./virtual-items"
+import { MessageComment, SummaryDiff, TimelineRow, TimelineRowMap } from "@opencode-ai/session-ui/v2/timeline-rows-v2"
 
 const emptyMessages: MessageType[] = []
 const emptyParts: PartType[] = []
