+++ packages/session-ui/src/v2/components/timeline-observe-element-offset-v2.ts
