+++ packages/session-ui/src/v2/components/timeline-playground-v2.stories.tsx
@@ -1,0 +1,13 @@
+import { TimelinePlayground } from "../../components/timeline-playground"
+
+export default {
+  title: "Playground/Timeline V2",
+  id: "playground-timeline-v2",
+  parameters: {
+    layout: "fullscreen",
+  },
+}
+
+export const Basic = {
+  render: () => <TimelinePlayground version="v2" />,
+}
